[UI구현] 강의 커뮤니티 / 강사
- 커뮤니티 게시글 작성
- 강사 상세 페이지
