+++ workspace/everyLearnPrj/src/main/webapp/WEB-INF/views/member/success_id_find.jsp
@@ -1,0 +1,46 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>Insert title here</title>
+<link href="/el/resources/css/member/success_id_find.css" rel="stylesheet" type="text/css">
+</head>
+<body>
+    <br><br><br><br><br>
+    <hr>
+    <div id="contain">
+        <!-- <div id="top-id_find"> -->
+            <div>
+                <div class="top-id_find"><img id="img-logo" src="/el/resources/img/logo/el.png" alt=""></div>
+            </div>
+
+            <div id="wrap-id_find">
+                
+                <form action="" id="id_find-form">
+                <h3>아이디 찾기</h3>    
+
+                <h4><label for="">에브리 회원님의 아이디는</label></h4>
+                <div class="id_find-box">
+                    <div id="decla-id">ayj7137</div>
+                </div>            
+                <h4><label for="">입니다.</label></h4>
+
+                <div id="btn-id_find">
+                    <button type="button" class="btnJoin">
+                        <div class="text-id_find" id="find-pw">비밀번호 찾기</div>
+                    </button>
+                    <button type="button" class="btnJoin">
+                        <div class="text-id_find" id="login">로그인</div>
+                    </button>
+                </div>
+                </form>
+                
+                
+            </div>
+
+    </div>
+
+</body>
+</html>
