+++ workspace/everyLearnPrj/src/main/webapp/WEB-INF/views/member/success_pwd_find.jsp
@@ -1,0 +1,47 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>Insert title here</title>
+<link href="/el/resources/css/member/success_pwd_find.css" rel="stylesheet" type="text/css">
+</head>
+<body>
+    <br><br><br><br><br>
+    <hr>
+    <div id="contain">
+        <!-- <div id="top-pw_find"> -->
+            <div>
+                <div class="top-pw_find"><img id="img-logo" src="/el/resources/img/logo/el.png" alt=""></div>
+            </div>
+
+            <div id="wrap-pw_find">
+                
+                <form action="" id="pw_find-form">
+                <h3>비밀번호 찾기</h3>    
+
+                <h4><label for="">에브리 회원님의 비밀번호</label></h4>
+                <div class="pw_find-box">
+                    <div id="decla-id">ayj7137</div>
+                </div>            
+                <!-- <h4><label for="">입니다.</label></h4> -->
+                <p style="text-align: center;">* 조회 후, 비밀번호 변경을 권장드립니다.</p>
+
+                <div id="btn-pw_find">
+                    <button type="button" class="btnJoin">
+                        <div class="text-pw_find" id="find-pw">비밀번호 변경</div>
+                    </button>
+                    <button type="button" class="btnJoin">
+                        <div class="text-pw_find" id="login">로그인</div>
+                    </button>
+                </div>
+                </form>
+                
+                
+            </div>
+
+    </div>
+
+</body>
+</html>
