+++ workspace/everyLearnPrj/src/main/webapp/resources/css/class_comm/comm_write.css
@@ -1,0 +1,10 @@
+#contain-write{
+    margin: auto;
+    width: 600px;
+    height: 1000px;
+}
+
+#btn-write{
+    background-color: #1187CF;
+}
+
