+++ workspace/everyLearnPrj/src/main/webapp/resources/css/member/success_pwd_find.css
@@ -1,0 +1,151 @@
+body{
+    padding: 0px;
+    margin: 0;
+}
+
+
+
+#contain{
+    width: 60vw;
+    height: 1500px;
+    margin: auto;
+
+}
+
+
+#top-pw_find{
+    /* display: flex; */
+    justify-content: center;
+    /* width: 500px; */
+}
+
+#img-logo{
+    width: 150px;
+    height: 150px;
+}
+
+.top-pw_find{
+
+    margin: auto;
+    width: 500px;
+    text-align: center;
+   
+}
+
+
+
+#wrap-pw_find{
+    width: 500px;
+    margin: auto;
+    display: flex;
+    justify-content: center;
+    border: 2px solid #1187CF;
+    border-radius: 30px;
+    padding-top: 40px;
+    padding-bottom: 50px;
+}
+
+h3{
+    text-align: center;
+    margin-top: 0px;
+    margin-bottom: 30px;
+}
+
+h4{
+
+    width: 300px;
+    margin: auto;
+    margin-bottom: 5px;
+    margin-top: 10px;
+    text-align: center;
+}   
+
+#pw_find-form{
+    width:400px;
+    margin: auto;
+}
+
+input{
+    border: none;
+    width: 80%;
+}
+
+input:focus {outline:none;}
+
+.pw_find-box{
+    width: 350px;
+    height: 50px;
+    border: solid 2px #dadada;
+    padding: 10px 10px 10px 10px;
+    background: #fff;
+    box-sizing: border-box;
+    margin: auto;
+    display: flex;
+    justify-self: center;
+    border-radius: 5px;
+    text-align: center;
+    margin-top: 10px;
+}
+
+#title-pw_find{
+    font-size: large;
+    font-weight: 900;
+}
+
+#decla-id{
+    text-align: center;
+    width: 350px;
+    height: 30px;
+}
+
+.btnJoin{
+    margin: auto;
+    width: 170px;
+    height: 50px;
+    margin-top: 10px;
+    display: flex;
+    justify-self: center;
+    text-align: center;
+    border: none;
+    background-color: #1187CF;
+    border-radius: 5px;
+}
+
+.text-pw_find{
+    margin: auto;
+    background-color: #1187CF;
+    font-size: large;
+    color: white;
+    font-weight: 700;
+    border: 1px solid #1187CF;
+}
+
+/* ---------- */
+#double-check{
+    font-size: 11px;
+    background-color: #1187CF;
+    color: white;
+    border: none;
+    border-radius: 5px;
+}
+
+#find{
+    width: 500px;
+    margin: auto;
+    text-align: center;
+    margin-top: 7px;
+}
+
+#find>a{
+    text-decoration: none;
+    font-size: small;
+    font-weight: 500;
+    color: gray;
+}
+
+#btn-pw_find{
+    margin: auto;
+    width: 350px;
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+}
